Create a file for interfaces
