+++ src/features/coins/coinsSlice.ts
@@ -1,46 +1,6 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import type { PayloadAction } from "@reduxjs/toolkit";
-export interface coinProp {
-  id: string,
-  name: string,
-  iconUrl: string,
-  marketCap: string,
-  price: string,
-  rank: number,
-  change: string,
-}
-interface coinsProp {
-  status: string,
-  data: {stats: globalStatsProp, coins: coinProp[]}
-}
-interface globalStatsProp {
-  totalCoins: number,
-  totalMarketCap: string,
-  total24hVolume: string,
-  totalExchanges: number,
-  totalMarkets: number,
-}
-
-interface singleCoinProp extends coinProp {
-  symbol: string,
-  description: string,
-  "24hVolume": string,
-  allTimeHigh: { price: string, timeStamp:number },
-  links: {name: string, type: string, url: string}[],
-  supply: { total: string, circulating: string},
-  numberOfMarkets: number,
-  numberOfExchanges: number,
-
-}
-interface initialStateProp {
-  loading: boolean,
-  coins: coinProp[],
-  singleCoin: singleCoinProp,
-  error: string | undefined,
-  globalStats: globalStatsProp,
-}
-
-// https://coinranking1.p.rapidapi.com/coins?limit=50
+import { initialStateProp, globalStatsProp, singleCoinProp, coinsProp } from "../../interface";
 const baserUrl = 'https://coinranking1.p.rapidapi.com/'
 const options = {
 	method: 'GET',
